CQT processing and plots
Added process_cqts.py that takes the MIDI and mp3 files and computes the
cqts as well as a distance matrix for every song in songlist.

plot_cqts.py creates 3 plot per song: One with the MIDI CQT, one with
the mp3 CQT and one with the distance matrix.

The resulting plots look somewhat like the plots in the paper, though
apparently the position of the sample audio has changed. resulting in
different timing.
Additionally there is still an error in the plots of the cqts since the
starting note (given in feature_extraction.py in the midi-dataset) seems
to be wrong.
Colors my have to be changed as well.